Fix a SQLAlchemy exception.
SQLAlchemy raises exceptions saying "ProgrammingError: SQLite objects
created in a thread can only be used in that same thread.The object
was created in thread ...". Some investigation indicates that this
involves the AMsoil configdb, which is a SQLite database.

Add the 'check_same_thread': False connection argument when creating
the SQLAlchemy engine to avoid the issue. We think this is ok because
the configdb is a readonly database for chapi.